feat: extract voxel instance struct
